Rename: group api params 수정
